refactor: useCoupons 파일 이름을 컨벤션에 맞게 수정
